linux/user_ptr.h: clear up as_user_ptr() usage
as_user_ptr() is intended to be used where an arbitrary integer e.g. an
error code is stored in a __user ptr.

The current implementation can be somewhat confusing in that it looks
like it can be used as direct replacement for u64_to_user_ptr e.g. in
PCuABI, where u64 addresses in the kernel-user interface are being
replaced with capability containing types such as __kernel_uintptr_t.

Currently, passing a valid capability e.g. a __kernel_uintptr_t,
__uintcap_t or user_uintptr_t etc to as_user_ptr() will result in a cast
to (void __user *) and a valid capability/pointer that can be
dereferenced. This is contrary to the code comment and intended usage
for as_user_ptr().

Add a cast to (u64) before the cast to (void __user *)(user_uintptr_t)
to make clearer the intended usage. This also always results in a null
capability that cannot be dereferenced.

(cherry picked from commit 1e06da936b0389360beb8b0552fe306d77360456)

+++ include/linux/user_ptr.h
@@ -12,10 +12,10 @@
  * as_user_ptr - convert an arbitrary integer value to a user pointer
  * @x: the integer value to convert
  *
- * Returns @x represented as a user pointer. The result is not a valid pointer
- * and shall not be dereferenced.
+ * Returns up to 64 bits of @x represented as a user pointer. The result is
+ * not a valid pointer and shall not be dereferenced.
  */
-#define as_user_ptr(x) ((void __user *)(user_uintptr_t)(x))
+#define as_user_ptr(x) ((void __user *)(user_uintptr_t)(u64)(x))
 
 /* Same semantics as as_user_ptr(), but also requires x to be of a given type */
 #define as_user_ptr_strict(type, x) (	\
